Move test helper functions to own file
